frontend/requirements.txt
streamlit
requests
python-jose